Separate new mandat views
